svg/geometry/parsing/rx-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="8000"><title>SVG Geometry Properties: parsing rx with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/geometry.html#RX" rel="help"/><h:meta content="rx supports only the grammar '&lt;length-percentage&gt; | auto'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("rx", "10");
test_invalid_value("rx", "none");
test_invalid_value("rx", "10px 20px");
test_invalid_value("rx", "-1px");

  ]]></script></svg>
